Move method signature to types
