+++ tempo/serve/pipeline.py
@@ -2,7 +2,7 @@ from types import SimpleNamespace
 from typing import Any, Callable, Optional
 
 from tempo.serve.base import BaseModel
-from tempo.serve.constants import ModelDataType
+from tempo.serve.types import ModelDataType
 from tempo.serve.metadata import ModelFramework, RuntimeOptions
 from tempo.serve.model import Model
 from tempo.serve.protocol import Protocol
